+++ common/src/main/java/net/mehvahdjukaar/polytone/slotify/BlitModifier.java
@@ -3,6 +3,7 @@ package net.mehvahdjukaar.polytone.slotify;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.mehvahdjukaar.polytone.utils.StrOpt;
+import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.renderer.texture.TextureAtlasSprite;
 import net.minecraft.resources.ResourceLocation;